Preparing codebase
