+++ data/net_arch_performance.csv
@@ -1,0 +1,201 @@
+default,128,256,64_32,64_64,128_64,256_128
+7,18,22,1,5,0,19
+8,2,5,4,12,9,15
+9,20,19,18,1,16,15
+7,5,10,3,14,13,5
+23,9,5,5,13,3,13
+17,21,7,4,18,3,21
+20,2,8,16,1,14,14
+1,1,15,7,16,8,15
+1,11,18,2,9,1,8
+19,4,5,9,14,7,16
+17,13,15,10,1,7,19
+8,1,15,8,9,13,14
+18,15,14,15,9,0,16
+2,15,12,11,19,13,8
+9,22,5,1,12,3,2
+22,3,13,12,2,10,8
+4,1,5,16,11,15,6
+15,8,17,17,7,5,19
+0,9,9,6,17,0,6
+3,1,15,15,1,9,13
+19,6,2,9,11,20,14
+17,11,21,13,1,1,15
+9,15,13,17,0,0,16
+10,16,1,7,4,1,8
+8,12,14,6,1,11,8
+13,1,10,3,12,11,2
+19,11,12,13,5,11,15
+20,12,7,5,7,11,14
+19,9,20,16,15,16,6
+16,17,6,7,11,2,13
+18,12,19,1,21,15,9
+4,4,5,11,11,19,16
+20,12,0,19,5,6,15
+5,14,16,15,13,2,2
+0,14,12,10,9,1,17
+18,7,15,4,10,12,13
+12,15,9,7,18,4,19
+23,7,10,7,9,2,7
+1,10,10,15,9,3,13
+12,17,9,4,17,14,18
+12,11,17,13,10,20,15
+21,20,12,14,5,21,18
+8,11,6,13,18,18,6
+9,19,3,1,2,9,15
+16,18,2,10,3,20,11
+7,9,6,9,13,5,6
+19,9,9,12,16,13,19
+14,19,10,1,0,11,11
+25,16,12,19,7,17,14
+18,16,13,14,12,5,15
+20,9,10,1,18,5,14
+19,2,9,6,9,2,19
+13,5,1,12,15,10,24
+23,3,17,16,6,11,13
+15,10,6,11,5,15,10
+0,7,12,4,5,16,7
+8,14,13,4,14,11,2
+4,2,22,12,11,9,16
+11,9,17,22,6,1,5
+15,17,13,21,19,5,14
+1,12,14,18,1,1,4
+16,16,13,13,21,7,5
+15,6,17,21,13,4,3
+9,11,15,11,16,2,2
+12,14,17,14,1,2,17
+10,17,1,18,14,15,2
+4,16,10,3,9,1,8
+7,11,8,3,16,6,18
+9,8,8,13,13,3,3
+5,11,15,16,17,18,15
+9,2,15,20,14,13,20
+15,11,14,1,17,11,19
+15,5,14,13,10,0,3
+10,10,15,3,3,6,7
+20,7,14,3,11,2,7
+3,5,15,1,15,13,15
+7,8,14,7,17,17,22
+15,3,10,20,6,2,16
+2,10,1,5,21,2,7
+21,11,6,5,1,17,9
+4,17,14,3,2,14,7
+3,12,3,15,9,12,15
+13,14,0,1,11,12,11
+6,14,3,13,13,1,18
+13,16,2,0,2,11,9
+16,18,17,14,12,3,12
+11,5,19,3,10,14,1
+3,1,12,16,6,7,14
+3,11,15,4,8,18,14
+16,14,17,1,13,7,18
+21,21,11,7,11,2,16
+21,15,15,13,3,11,17
+16,15,13,6,9,17,6
+18,5,21,3,1,19,13
+14,16,14,15,3,12,22
+3,12,21,19,14,15,19
+0,1,16,3,1,10,15
+4,3,16,10,18,6,23
+4,4,13,0,15,0,16
+19,15,14,4,1,9,12
+10,4,4,1,8,15,22
+15,3,1,5,15,13,16
+23,5,10,16,14,5,15
+12,9,7,3,1,11,10
+17,3,3,6,1,3,7
+13,17,0,19,9,20,17
+4,11,12,14,14,3,11
+15,5,7,18,17,13,7
+5,15,14,8,14,1,9
+14,17,15,16,7,3,12
+19,15,7,2,16,16,5
+21,6,15,11,8,1,15
+6,14,4,5,14,0,15
+5,1,10,11,12,5,19
+6,4,5,4,8,16,7
+9,13,19,6,12,16,8
+22,4,11,15,10,13,9
+9,1,19,14,14,5,17
+21,5,14,1,13,1,17
+9,12,12,6,11,9,19
+12,17,6,6,9,1,16
+17,9,17,15,2,0,12
+3,2,15,6,6,19,17
+1,7,18,11,14,5,15
+20,4,3,17,1,15,15
+0,15,9,8,6,2,8
+4,11,15,11,11,1,18
+5,1,12,3,1,17,15
+13,12,15,15,16,10,2
+13,20,11,9,4,7,20
+2,18,19,0,10,11,12
+3,14,14,17,9,17,18
+10,18,3,2,10,15,16
+2,19,3,14,7,10,3
+3,1,18,20,6,15,14
+5,2,4,2,15,15,19
+18,14,6,15,8,13,3
+18,2,13,17,9,6,12
+19,15,0,18,15,1,15
+11,3,16,17,15,2,20
+18,16,14,4,1,12,18
+18,1,3,4,19,0,22
+9,16,17,6,15,2,1
+1,3,18,15,2,4,9
+9,2,7,18,10,7,15
+13,2,21,15,12,7,13
+19,1,0,5,2,1,20
+3,10,6,22,9,13,3
+14,3,1,4,11,1,16
+20,2,9,5,7,17,6
+6,15,20,1,14,16,24
+5,3,14,15,12,12,15
+19,14,4,6,19,3,14
+17,5,10,8,6,2,18
+9,15,11,15,0,22,15
+3,1,13,1,11,1,13
+6,14,5,11,8,0,17
+0,16,11,17,10,15,14
+5,2,1,0,7,6,0
+20,18,3,13,13,0,8
+14,12,16,11,1,15,17
+1,4,12,0,15,18,2
+6,3,4,9,17,1,8
+17,14,13,18,18,3,3
+4,7,8,5,15,1,8
+0,13,18,3,11,4,13
+12,2,7,5,12,4,20
+14,17,17,0,11,5,12
+24,14,10,12,0,20,18
+12,2,11,7,7,5,10
+13,9,3,19,9,17,9
+1,3,9,12,14,2,9
+13,10,17,9,5,4,17
+5,18,12,11,13,1,2
+20,16,15,11,10,0,14
+13,12,12,4,14,2,18
+0,8,14,0,3,3,3
+4,2,3,8,16,21,17
+10,10,15,3,4,20,7
+18,2,18,15,17,11,12
+15,13,15,16,8,15,14
+14,23,1,6,1,4,14
+19,7,19,16,1,1,17
+13,11,7,13,24,21,14
+19,1,3,12,3,4,8
+9,11,21,2,8,17,15
+17,8,18,19,7,17,14
+2,5,13,16,0,6,1
+3,3,4,0,14,5,16
+19,17,18,5,4,18,15
+5,6,13,6,0,17,15
+0,11,20,13,3,11,8
+3,8,3,10,15,13,14
+15,14,15,6,6,1,14
+4,18,10,7,13,7,3
+1,10,15,14,11,14,5
+16,1,19,14,7,10,20
+21,12,0,14,11,19,11
+5,5,15,5,10,17,13
+3,15,14,14,3,2,16
