+++ pyproject.toml
@@ -10,6 +10,9 @@ python = "^3.9"
 stable-baselines3 = "^2.4.0"
 torch = "2.2.2"
 ipykernel = "^6.29.5"
+imitation = "^1.0.0"
+plotly = "^5.24.1"
+nbformat = "^5.10.4"
 
 [build-system]
 requires = ["poetry-core"]
